+++ lib/ProductOpener/Config_off.pm
@@ -67,6 +67,9 @@ BEGIN {
 		$folksonomy_url
 		$process_global_redis_events
 
+		$recipe_estimator_url
+		$recipe_estimator_scipy_url
+
 		$mongodb
 		$mongodb_host
 		$mongodb_timeout_ms
@@ -439,6 +442,14 @@ $crowdin_project_key = $ProductOpener::Config2::crowdin_project_key;
 # enable an in-site robotoff-asker in the product page
 $robotoff_url = $ProductOpener::Config2::robotoff_url;
 $query_url = $ProductOpener::Config2::query_url;
+
+# recipe-estimator product service
+# To test a locally running recipe-estimator with product opener in a docker dev environment:
+# - run recipe-estimator with `uvicorn recipe_estimator.main:app --reload --host 0.0.0.0`
+# $recipe_estimator_url = "http://host.docker.internal:8000/api/v3/estimate_recipe";
+
+$recipe_estimator_url = $ProductOpener::Config2::recipe_estimator_url;
+$recipe_estimator_scipy_url = $ProductOpener::Config2::recipe_estimator_scipy_url;
 
 # do we want to send emails
 $log_emails = $ProductOpener::Config2::log_emails;